requirements.txt
asgiref==3.4.1
Django==3.2.8
pytz==2021.3
sqlparse==0.4.2
django-environ
requests~=2.28.0